Image-to-text using pseudo label
Deep learning algorithms are being used to detect and recognize car license plate for intelligent vehicle control systems. License plate recognition goes through the stages of object detection and character recognition. Character recognition is mainly trained using supervised learning method while it consumes a lot of cost in labeling the collected dataset. In this paper, we propose a license plate detection and recognition model using a semi supervised learning method. The proposed model consists of a VGG 16 based
detector and a CNN and BiLSTM based recognizer. The proposed model is an end to end model that dose not require third party modules such as OCR. The proposed model utilizes unlabeled data and shows higher performance than when only labeled data is used. The proposed model can also be used for other semi structured documents and function as a labler for large scale data.